# openapi.yaml
openapi: 3.1.0
info:
  title: Human Design API
  version: 3.4.1
paths:
...
  /calculate:
    get:
      tags:
      - general
      summary: Calculate Hd
      operationId: calculate_hd_calculate_get
      security:
      - HTTPBearer: []
      parameters:
      - name: year
        in: query
        required: false
        schema:
          type: integer
          description: Birth year
          default: 1968
          title: Year
        description: Birth year
      - name: month
        in: query
        required: false
        schema:
          type: integer
          description: Birth month
          default: 2
          title: Month
        description: Birth month
      - name: day
        in: query
        required: false
        schema:
          type: integer
          description: Birth day
          default: 21
          title: Day
        description: Birth day
      - name: hour
        in: query
        required: false
        schema:
          type: integer
          description: Birth hour
          default: 11
          title: Hour
        description: Birth hour
      - name: minute
        in: query
        required: false
        schema:
          type: integer
          description: Birth minute
          default: 0
          title: Minute
        description: Birth minute
      - name: second
        in: query
        required: false
        schema:
          type: integer
          description: Birth second (optional, default 0)
          default: 0
          title: Second
        description: Birth second (optional, default 0)
      - name: place
        in: query
        required: false
        schema:
          type: string
          description: Birth place (city, country)
          default: Kirikkale, Turkey
          title: Place
        description: Birth place (city, country)
      - name: gender
        in: query
        required: false
        schema:
          type: string
          description: Gender (optional)
          default: male
          title: Gender
        description: Gender (optional)
      - name: islive
        in: query
        required: false
        schema:
          type: boolean
          description: Whether the person is still alive (True) or deceased (False)
          default: true
          title: Islive
        description: Whether the person is still alive (True) or deceased (False)
      - name: latitude
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: 'null'
          description: Optional latitude for birth place
          title: Latitude
        description: Optional latitude for birth place
      - name: longitude
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: 'null'
          description: Optional longitude for birth place
          title: Longitude
        description: Optional longitude for birth place
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v2/calculate:
    post:
      tags:
      - v2
      summary: Calculate Hd V2
      operationId: calculate_hd_v2_v2_calculate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculateRequestV2'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculateResponseV2'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
    AdvancedSectionV2:
      properties:
        dream_rave:
          anyOf:
          - $ref: '#/components/schemas/DreamRaveOutput'
          - type: 'null'
        global_cycle:
          anyOf:
          - $ref: '#/components/schemas/GlobalCycleOutput'
          - type: 'null'
      type: object
      title: AdvancedSectionV2
    AnalyticsSectionV2:
      properties:
        motivation:
          anyOf:
          - type: string
          - type: 'null'
          title: Motivation
        perspective:
          anyOf:
          - type: string
          - type: 'null'
          title: Perspective
        quarter:
          anyOf:
          - $ref: '#/components/schemas/QuarterV2'
          - type: 'null'
        line_counts:
          anyOf:
          - $ref: '#/components/schemas/LineCountsV2'
          - type: 'null'
        sun_roles:
          anyOf:
          - $ref: '#/components/schemas/SunRolesV2'
          - type: 'null'
        yin_yang_balance:
          anyOf:
          - $ref: '#/components/schemas/YinYangBalanceV2'
          - type: 'null'
        contour:
          anyOf:
          - $ref: '#/components/schemas/ContourV2'
          - type: 'null'
      type: object
      title: AnalyticsSectionV2
      description: 'Extended analytics: quarter, line counts, sun roles, yin/yang, contour.'
...
    CalculateRequestV2:
      properties:
        year:
          type: integer
          title: Year
          description: Birth year
          default: 1968
        month:
          type: integer
          title: Month
          description: Birth month
          default: 2
        day:
          type: integer
          title: Day
          description: Birth day
          default: 21
        hour:
          type: integer
          title: Hour
          description: Birth hour
          default: 11
        minute:
          type: integer
          title: Minute
          description: Birth minute
          default: 0
        second:
          type: integer
          title: Second
          description: Birth second
          default: 0
        place:
          type: string
          title: Place
          description: Birth place
          default: Kirikkale, Turkey
        gender:
          anyOf:
          - type: string
          - type: 'null'
          title: Gender
          description: Gender
          default: male
        islive:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Islive
          description: Whether alive
          default: true
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        include:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Include
          description: Sections to include (e.g. ['general', 'personality_gates'])
          example:
          - general
          - personality_gates
        exclude:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Exclude
          description: Sections to exclude
          example:
          - channels
      type: object
      title: CalculateRequestV2
    CalculateResponseV2:
      properties:
        general:
          anyOf:
          - $ref: '#/components/schemas/GeneralSectionV2'
          - type: 'null'
        centers:
          anyOf:
          - $ref: '#/components/schemas/CentersV2'
          - type: 'null'
        channels:
          anyOf:
          - items:
              type: object
            type: array
          - type: 'null'
          title: Channels
        variables:
          anyOf:
          - $ref: '#/components/schemas/VariablesV2'
          - type: 'null'
        gates:
          anyOf:
          - $ref: '#/components/schemas/GatesV2'
          - type: 'null'
        mechanics:
          anyOf:
          - type: object
          - type: 'null'
          title: Mechanics
        analytics:
          anyOf:
          - $ref: '#/components/schemas/AnalyticsSectionV2'
          - type: 'null'
        advanced:
          anyOf:
          - $ref: '#/components/schemas/AdvancedSectionV2'
          - type: 'null'
      type: object
      title: CalculateResponseV2
    CentersV2:
      properties:
        defined:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Defined
        undefined:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Undefined
      type: object
      title: CentersV2
    ClassBreakdownV2:
      properties:
        values:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Values
        share_pct:
          anyOf:
          - type: integer
          - type: 'null'
          title: Share Pct
        total:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total
      type: object
      title: ClassBreakdownV2
      description: Classification result with planet-gate values and share percentage.
    ContourV2:
      properties:
        realization:
          anyOf:
          - additionalProperties:
              $ref: '#/components/schemas/ClassBreakdownV2'
            type: object
          - type: 'null'
          title: Realization
        mind:
          anyOf:
          - additionalProperties:
              $ref: '#/components/schemas/ClassBreakdownV2'
            type: object
          - type: 'null'
          title: Mind
        decision:
          anyOf:
          - additionalProperties:
              $ref: '#/components/schemas/ClassBreakdownV2'
            type: object
          - type: 'null'
          title: Decision
        big_o:
          anyOf:
          - additionalProperties:
              $ref: '#/components/schemas/ClassBreakdownV2'
            type: object
          - type: 'null'
          title: Big O
      type: object
      title: ContourV2
      description: 'Contour analysis: realization, mind, decision, big_o classifications.'
    DreamRaveOutput:
      properties:
        activated_centers:
          items:
            type: string
          type: array
          title: Activated Centers
        activated_gates:
          items:
            type: integer
          type: array
          title: Activated Gates
        status:
          type: string
          title: Status
      type: object
      required:
      - activated_centers
      - activated_gates
      - status
      title: DreamRaveOutput
    GateV2:
      properties:
        gate:
          type: integer
          title: Gate
        line:
          type: integer
          title: Line
        color:
          type: integer
          title: Color
        tone:
          type: integer
          title: Tone
        base:
          type: integer
          title: Base
        lon:
          type: number
          title: Lon
        gate_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Gate Name
        gate_summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Gate Summary
        line_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Line Name
        line_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Line Description
        fixation:
          anyOf:
          - type: object
          - type: 'null'
          title: Fixation
      type: object
      required:
      - gate
      - line
      - color
      - tone
      - base
      - lon
      title: GateV2
    GatesV2:
      properties:
        personality:
          anyOf:
          - additionalProperties:
              $ref: '#/components/schemas/GateV2'
            type: object
          - type: 'null'
          title: Personality
        design:
          anyOf:
          - additionalProperties:
              $ref: '#/components/schemas/GateV2'
            type: object
          - type: 'null'
          title: Design
      type: object
      title: GatesV2
    GeneralSectionV2:
      properties:
        birth_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Birth Date
        create_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Create Date
        birth_place:
          anyOf:
          - type: string
          - type: 'null'
          title: Birth Place
        age:
          anyOf:
          - type: integer
          - type: 'null'
          title: Age
        gender:
          anyOf:
          - type: string
          - type: 'null'
          title: Gender
        islive:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Islive
        zodiac_sign:
          anyOf:
          - type: string
          - type: 'null'
          title: Zodiac Sign
        energy_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Energy Type
        strategy:
          anyOf:
          - type: string
          - type: 'null'
          title: Strategy
        signature:
          anyOf:
          - type: string
          - type: 'null'
          title: Signature
        not_self:
          anyOf:
          - type: string
          - type: 'null'
          title: Not Self
        aura:
          anyOf:
          - type: string
          - type: 'null'
          title: Aura
        inner_authority:
          anyOf:
          - type: string
          - type: 'null'
          title: Inner Authority
        inc_cross:
          anyOf:
          - type: string
          - type: 'null'
          title: Inc Cross
        profile:
          anyOf:
          - type: string
          - type: 'null'
          title: Profile
        definition:
          anyOf:
          - type: string
          - type: 'null'
          title: Definition
      type: object
      title: GeneralSectionV2
    GlobalCycleOutput:
      properties:
        great_cycle:
          type: string
          title: Great Cycle
        cycle_cross:
          type: string
          title: Cycle Cross
        gates:
          items:
            type: integer
          type: array
          title: Gates
        description:
          type: string
          title: Description
      type: object
      required:
      - great_cycle
      - cycle_cross
      - gates
      - description
      title: GlobalCycleOutput
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LineCountsV2:
      properties:
        prs:
          anyOf:
          - additionalProperties:
              type: integer
            type: object
          - type: 'null'
          title: Prs
        des:
          anyOf:
          - additionalProperties:
              type: integer
            type: object
          - type: 'null'
          title: Des
        total:
          anyOf:
          - additionalProperties:
              type: integer
            type: object
          - type: 'null'
          title: Total
      type: object
      title: LineCountsV2
    QuarterV2:
      properties:
        number:
          anyOf:
          - type: integer
          - type: 'null'
          title: Number
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      type: object
      title: QuarterV2
...
    SunRoleItemV2:
      properties:
        hexagram:
          anyOf:
          - type: string
          - type: 'null'
          title: Hexagram
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
      type: object
      title: SunRoleItemV2
    SunRolesV2:
      properties:
        prs_sun:
          anyOf:
          - $ref: '#/components/schemas/SunRoleItemV2'
          - type: 'null'
        des_sun:
          anyOf:
          - $ref: '#/components/schemas/SunRoleItemV2'
          - type: 'null'
      type: object
      title: SunRolesV2
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VariableItemV2:
      properties:
        value:
          anyOf:
          - type: string
          - type: 'null'
          title: Value
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        aspect:
          anyOf:
          - type: string
          - type: 'null'
          title: Aspect
        def_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Def Type
      type: object
      title: VariableItemV2
    VariablesV2:
      properties:
        top_right:
          anyOf:
          - $ref: '#/components/schemas/VariableItemV2'
          - type: 'null'
        bottom_right:
          anyOf:
          - $ref: '#/components/schemas/VariableItemV2'
          - type: 'null'
        top_left:
          anyOf:
          - $ref: '#/components/schemas/VariableItemV2'
          - type: 'null'
        bottom_left:
          anyOf:
          - $ref: '#/components/schemas/VariableItemV2'
          - type: 'null'
        short_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Short Code
      type: object
      title: VariablesV2
    YinYangBalanceV2:
      properties:
        Yang:
          anyOf:
          - $ref: '#/components/schemas/ClassBreakdownV2'
          - type: 'null'
        Yin:
          anyOf:
          - $ref: '#/components/schemas/ClassBreakdownV2'
          - type: 'null'
        Balance:
          anyOf:
          - $ref: '#/components/schemas/ClassBreakdownV2'
          - type: 'null'
      type: object
      title: YinYangBalanceV2
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer